:art:更新user模块实体类

+++ felix-system/felix-user/src/main/java/com/mazhj/felix/user/pojo/model/Bookshelf.java
@@ -13,9 +13,11 @@ public class Bookshelf implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private Integer id;
+    private Long id;
 
+    /** 用户id*/
     private String userId;
 
+    /** 图书id*/
     private String bookId;
 }
